update 1.01
Changing the mod NavInstrument installation in my MFD, there is no possibility for the moment to fit the background Handler between my top and lower page info key, the mfd HSI and AI are partly hidden by my page text, message text "no software installed" not functional with NavInstruments, i'm using ModuleManager in a same way of original mod to keep compatibility.
Finished the startup page for MFD60*30, corrected the label name in cockpit.
and that will be all for now.